+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.2
 
 - add two js entry module support: `test/inner.js`, `test/inner/index.js` (module: `test/inner`)
+- fixed fatal bugs by add `local lila`
 
 ## 0.0.1